+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -105,6 +105,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Light decomposition lighting debugging modes and support in AOV
 - Added exposure compensation to Fixed exposure mode
 - Added support for rasterized area light shadows in StackLit
+- Added alpha channel support in motion blur pass.
+- Added the HDRP Compositor Tool (in Preview).
 
 ### Fixed
 - Fix when rescale probe all direction below zero (1219246)